content/canvas/test/test_2d.imageData.create.round.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.create.round - bug 433004</title>
<!-- Testing: createImageData(w, h) is rounded the same as getImageData(0, 0, w, h) -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;

var imgdata1 = ctx.createImageData(10.01, 10.99);
var imgdata2 = ctx.getImageData(0, 0, 10.01, 10.99);
ok(imgdata1.width == imgdata2.width, "imgdata1.width == imgdata2.width");
ok(imgdata1.height == imgdata2.height, "imgdata1.height == imgdata2.height");

SimpleTest.finish();

});
</script>
